wip: add main prover backend traits
