+++ Cargo.toml
@@ -86,7 +86,7 @@ tracing = "0.1.40"
 serde_json = "1.0.117"
 lazy_static = "1.5.0"
 once_cell = "1.19.0"
-derive-new = "0.6.0"
+derive-new = "0.7.0"
 derive_more = "1.0.0"
 derivative = "2.2.0"
 strum_macros = "0.26.4"
