+++ crates/stark-backend/src/lib.rs
@@ -31,6 +31,8 @@ pub mod interaction;
 pub mod keygen;
 /// Polynomials
 pub mod poly;
+/// Definition of the STARK proof struct.
+pub mod proof;
 /// Prover implementation for partitioned multi-matrix AIRs.
 pub mod prover;
 /// Trait for RAP (Randomized AIR with Preprocessing)
