+++ crates/stark-backend/src/prover/quotient/helper.rs
@@ -1,6 +1,5 @@
 use crate::{
     config::StarkGenericConfig,
-    interaction::HasInteractionChunkSize,
     keygen::types::{MultiStarkProvingKey, StarkProvingKey},
     prover::quotient::QuotientVkData,
 };
@@ -13,8 +12,6 @@ impl<SC: StarkGenericConfig> QuotientVkDataHelper<SC> for StarkProvingKey<SC> {
     fn get_quotient_vk_data(&self) -> QuotientVkData<SC> {
         QuotientVkData {
             quotient_degree: self.vk.quotient_degree,
-            rap_phase_seq_kind: self.vk.rap_phase_seq_kind,
-            interaction_chunk_size: self.rap_phase_seq_pk.interaction_chunk_size(),
             symbolic_constraints: &self.vk.symbolic_constraints,
         }
     }
